pyproject.toml
[project]
name = "nutriontracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.6",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
